fix #76001 delete clef at start of measure before adding new clef
If user drags a clef onto a measure, then Score::undoChangeClef will be called.  The old behavior did not remove any clef at beginning of the measure if the segment to add to was at start of measure, before actually adding the clef to the end of the previous measure.  I've fixed this case by first checking for existing clef at start of measure and deleting it before proceeding with the rest of the function which adds the new clef to end of previous measure (or modifies current one if exists).  (Note: Segment::remove(Clef*) already check to make sure clef is not a "generated" clef, so unnecessary for me to add that test.)