fix #76006 insert clef to meas after section break will insert at meas start
Previously, if insert clef to a measure, then clef will be inserted to end of previous measure, except for case that is start of score in which case the new clef is added to beginning of the measure.  This commit changes behavior so that if measure is at start of section, then the new clef will be inserted at beginning of measure, for same behavior as if was at beginning of score.