fix #202751 SegFault delete SFont while loadSamples()
Mutex may be locked if another thread is trying to load, add, or delete soundfonts.

If another thread deletes a soundfont, trying to read the zone info will result in a SEGFAULT as described in issue #202751.

These operations need to be mutually-exclusive, but since Seq::process() is a realtime context, we can't block here.

In order to keep Seq::process() running, my solution here is to simply not execute the rest of loadSamples() if another thread had locked that mutex.