fix #280364 measure start timewise input
Previous code crashed when doing a timewise insert at the beginning of a measure. This was because the call to msMeasure->findFirst searches for a segment at tick 0, but segment ticks had already been shifted forward in time with the earlier call to InsertTime. Using msMeasure->first(SegmentType::ChordRest) avoids that problem.

Second problem was a crash when inserting to beginning of a multi-measure rest.  This was fixed by testing for isMMRest() and using the multi-measure rest if true.

I also had to change behavior to convert all measure-rests into normal (non-measure) rests(s) of equivalent duration first before inserting any additional space.  Unfortunately it would have been ideal to simply keep measure rests and exapand them to fill the new space, but I couldn't figure out a clean way to do that.  So I opted for the simpler and safer method of converting them into...